install laravel breeze and implement proper authentication

+++ app/Http/Controllers/Api/V1/PostController.php
@@ -6,7 +6,7 @@ use App\Http\Controllers\Controller;
 use App\Http\Requests\StorePostRequest;
 use App\Http\Resources\PostResource;
 use App\Models\Post;
-use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 
 class PostController extends Controller
 {
@@ -15,7 +15,9 @@ class PostController extends Controller
      */
     public function index()
     {
-        return PostResource::collection(Post::with('author')->paginate(10));
+        $user = request()->user();
+        $posts = $user->posts()->paginate(10);
+        return PostResource::collection($posts);
     }
 
     /**
@@ -25,7 +27,7 @@ class PostController extends Controller
     {
         $data = $request->validated();
         // $data['author_id'] = $request->user()->id; 
-        $data['author_id'] = 1;
+        $data['author_id'] = $request->user()->id; // Set the author_id to the authenticated user's ID
         $post = Post::create($data);
         return new PostResource($post);
     }
@@ -36,7 +38,7 @@ class PostController extends Controller
     public function show(Post $post)
 
     {
-
+        abort_if($post->author_id !== Auth::id(), 403, 'Unauthorized access to this post.');
         return new PostResource($post);
     }
 
@@ -45,6 +47,7 @@ class PostController extends Controller
      */
     public function update(StorePostRequest $request, Post $post)
     {
+        abort_if($post->author_id !== Auth::id(), 403, 'Unauthorized access to this post.');
         $data = $request->validated();
         $post->update($data);
         return new PostResource($post);
@@ -55,6 +58,7 @@ class PostController extends Controller
      */
     public function destroy(Post $post)
     {
+        abort_if($post->author_id !== Auth::id(), 403, 'Unauthorized access to this post.');
         $post->delete();
         return response()->json(['message' => 'Post deleted successfully']);
     }
